requirements.txt
# Core Framework
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20

# LLM Integration
ollama>=0.1.7
langchain-ollama>=0.1.0

# Vector Database
pinecone
sentence-transformers>=2.2.2

# Document Processing
PyPDF2>=3.0.1
python-docx>=0.8.11
pdfplumber>=0.9.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
streamlit-extras>=0.3.0

# Additional utilities
typing-extensions>=4.5.0
requests>=2.28.0
pathlib>=1.0.1